Adde a decorator to overrider deviice time check in !RELEASE builds
