+++ src/xcode/ENA/ENA/Source/Services/Risk/Provider/Helper/AppConfiguration+Features.swift
@@ -13,33 +13,3 @@ extension SAP_Internal_V2_ApplicationConfigurationIOS {
 	}
 }
 
-protocol AppFeatureProviding {
-	func value(for appFeature: SAP_Internal_V2_ApplicationConfigurationIOS.AppFeature) -> Bool
-}
-
-class AppFeatureProvider: AppFeatureProviding {
-
-	// MARK: - Init
-	init(
-		appConfiguration: AppConfigurationProviding
-	) {
-		self.appConfiguration = appConfiguration
-	}
-
-	// MARK: - Protocol AppFeaturesProviding
-
-	func value(for appFeature: SAP_Internal_V2_ApplicationConfigurationIOS.AppFeature) -> Bool {
-		guard let configuration = appConfiguration?.currentAppConfig.value else {
-			return false
-		}
-		let feature = configuration.appFeatures.appFeatures.first {
-			$0.label == appFeature.rawValue
-		}
-		return feature?.value == 1
-	}
-
-	// MARK: - Private
-
-	private weak var appConfiguration: AppConfigurationProviding?
-
-}
