+++ src/xcode/ENA/ENA/Source/Services/Risk/Provider/Helper/AppConfigurationProviding.swift
@@ -22,7 +22,7 @@ protocol AppConfigurationProviding: AnyObject {
 	/// a publisher for the current app configuration
 	var currentAppConfig: CurrentValueSubject<SAP_Internal_V2_ApplicationConfigurationIOS, Never> { get }
 
-	var featureProvider: AppFeatureProvider { get }
+	var featureProvider: AppFeatureProviding { get }
 }
 
 /// Some requirements for app configuration handling
